add support for serializing explanations using save_explanation and load_explanation with sparse data
