+++ python/interpret_community/explanation/__init__.py
@@ -3,3 +3,7 @@
 # ---------------------------------------------------------
 
 """Defines the building blocks for explanations returned by explainers."""
+
+from .serialization import load_explanation, save_explanation
+
+__all__ = ['load_explanation', 'save_explanation']
