+++ python/interpret_community/explanation/explanation.py
@@ -5,8 +5,6 @@
 """Defines the explanations that are returned from explaining models."""
 
 import gc
-import json
-import os
 import uuid
 from abc import ABC, abstractmethod
 
@@ -1769,6 +1767,9 @@ def _get_raw_explainer_create_explanation_kwargs(*, kwargs=None, explanation=Non
 def save_explanation(explanation, path, exist_ok=False):
     """Serialize the explanation.
 
+    Defined here for backwards compatibility.  The method has been moved to serialization.py.
+    Note this method is monkey patched there to avoid circular dependencies.
+
     :param explanation: The Explanation to be serialized.
     :type explanation: Explanation
     :param path: The path to the directory in which the explanation will be saved. By default, must be a new directory
@@ -1778,173 +1779,20 @@ def save_explanation(explanation, path, exist_ok=False):
         this function. If True, a prexisting path may be passed. Any preexisting files whose names match those of the
         files that make up the explanation will be overwritten.
     :type exist_ok: bool
-    :return: JSON-formatted explanation data.
-    :rtype: str
     """
-    if os.path.exists(path) and not exist_ok:
-        raise Exception('The directory specified by path already exists. '
-                        'Please pass in a new directory or set exists_ok=True.')
-    os.makedirs(path, exist_ok=True)
-
-    # TODO replace with set of params from below
-    uploadable_properties = [
-        ExplainParams.FEATURES,
-        ExplainParams.LOCAL_IMPORTANCE_VALUES,
-        ExplainParams.EXPECTED_VALUES,
-        ExplainParams.CLASSES,
-        ExplainParams.GLOBAL_IMPORTANCE_NAMES,
-        ExplainParams.GLOBAL_IMPORTANCE_RANK,
-        ExplainParams.GLOBAL_IMPORTANCE_VALUES,
-        ExplainParams.PER_CLASS_NAMES,
-        ExplainParams.PER_CLASS_RANK,
-        ExplainParams.PER_CLASS_VALUES,
-        ExplainParams.INIT_DATA,
-        ExplainParams.EVAL_DATA,
-        ExplainParams.EVAL_Y_PRED,
-        ExplainParams.EVAL_Y_PRED_PROBA
-    ]
-    # TODO will need to add viz data on top of this
-    for prop in uploadable_properties:
-        if hasattr(explanation, prop) and getattr(explanation, prop) is not None:
-            value = getattr(explanation, prop)
-            if isinstance(value, pd.DataFrame):
-                value = value.values.tolist()
-                metadata = 'DataFrame'
-            elif isinstance(value, DatasetWrapper):
-                value = value.original_dataset.tolist()
-                metadata = 'DatasetWrapper'
-            elif isinstance(value, np.ndarray):
-                value = value.tolist()
-                metadata = 'ndarray'
-            elif isinstance(value, list):
-                metadata = 'list'
-            else:
-                metadata = 'other'
-            data_dict = {
-                'metadata': metadata,
-                'data': value
-            }
-            filename = os.path.join(path, prop + '.json')
-            with open(filename, 'w') as f:
-                json.dump(data_dict, f)
-    # create metadata file
-    prop_dict = _get_explanation_metadata(explanation)
-
-    filename = os.path.join(path, 'explanation_metadata.json')
-    with open(filename, 'w') as f:
-        json.dump(prop_dict, f)
-
-
-def _get_explanation_metadata(explanation):
-    classification = ClassesMixin._does_quack(explanation)
-    is_raw = False if not FeatureImportanceExplanation._does_quack(explanation) else explanation.is_raw
-    is_eng = False if not FeatureImportanceExplanation._does_quack(explanation) else explanation.is_engineered
-    num_features = 0 if not FeatureImportanceExplanation._does_quack(explanation) else explanation.num_features
-    prop_dict = {
-        ExplainParams.EXPLANATION_ID: explanation.id,
-        ExplainType.MODEL: ExplainType.CLASSIFICATION if classification else ExplainType.REGRESSION,
-        ExplainType.DATA: ExplainType.TABULAR,
-        ExplainType.MODEL_TASK: explanation.model_task,
-        ExplainType.METHOD: explanation.method,
-        ExplainType.MODEL_CLASS: explanation.model_type,
-        ExplainType.IS_RAW: is_raw,
-        ExplainType.IS_ENG: is_eng,
-        ExplainType.GLOBAL: GlobalExplanation._does_quack(explanation),
-        ExplainType.LOCAL: LocalExplanation._does_quack(explanation),
-        ExplainParams.NUM_CLASSES: 1 if not ClassesMixin._does_quack(explanation) else explanation.num_classes,
-        ExplainParams.NUM_EXAMPLES: 0 if not LocalExplanation._does_quack(explanation) else explanation.num_examples,
-        ExplainParams.NUM_FEATURES: num_features,
-        'serialization_version': 1,
-    }
-    return prop_dict
-
-
-def _get_value_from_file(file_var):
-    json_input = json.load(file_var)
-    meta = json_input['metadata']
-    data = json_input['data']
-    if meta == 'DataFrame':
-        return pd.DataFrame(data)
-    elif meta == 'DatasetWrapper':
-        if isinstance(data, list):
-            data = np.array(data)
-        return DatasetWrapper(data)
-    elif meta == 'ndarray':
-        return np.array(data)
-    elif meta == 'list':
-        return data
-    else:
-        raise Exception('Unrecognized data type in deserialization: {}'.format(meta))
-
-
-def _get_kwargs(path, params, local_explanation=None):
-    numpy_params = [
-        ExplainParams.LOCAL_IMPORTANCE_VALUES,
-        ExplainParams.EVAL_DATA,
-        ExplainParams.INIT_DATA
-    ]
-    kwargs = {}
-    for param in params:
-        if os.path.exists(os.path.join(path, param + '.json')):
-            with open(os.path.join(path, param + '.json'), 'r') as f:
-                if param in numpy_params:
-                    kwargs[param] = np.array(_get_value_from_file(f))
-                else:
-                    kwargs[param] = _get_value_from_file(f)
-    with open(os.path.join(path, 'explanation_metadata.json'), 'r') as f:
-        metadata = json.load(f)
-        param_list = [
-            ExplainParams.METHOD,
-            ExplainParams.MODEL_TASK,
-            ExplainParams.NUM_FEATURES,
-            ExplainParams.IS_RAW,
-            ExplainParams.IS_ENG,
-            ExplainParams.EXPLANATION_ID
-        ]
-        for param in param_list:
-            kwargs[param] = metadata[param]
-        kwargs[ExplainParams.MODEL_TYPE] = metadata[ExplainType.MODEL_CLASS]
-        classification = metadata[ExplainType.MODEL] == ExplainType.CLASSIFICATION
-        kwargs[ExplainParams.CLASSIFICATION] = classification
-        if classification:
-            kwargs[ExplainParams.NUM_CLASSES] = metadata[ExplainParams.NUM_CLASSES]
-        if local_explanation is not None:
-            kwargs[ExplainParams.LOCAL_EXPLANATION] = local_explanation
-    return kwargs
+    pass
 
 
 def load_explanation(path):
-    shared_params = [
-        ExplainParams.EXPECTED_VALUES,
-        ExplainParams.FEATURES,
-        ExplainParams.CLASSES,
-        ExplainParams.EVAL_DATA,
-        ExplainParams.INIT_DATA,
-        ExplainParams.EVAL_Y_PRED,
-        ExplainParams.EVAL_Y_PRED_PROBA
-    ]
-    global_params = [
-        ExplainParams.GLOBAL_IMPORTANCE_RANK,
-        ExplainParams.GLOBAL_IMPORTANCE_VALUES,
-        ExplainParams.PER_CLASS_RANK,
-        ExplainParams.PER_CLASS_VALUES
-    ]
-    local_params = [
-        ExplainParams.LOCAL_IMPORTANCE_VALUES
-    ]
-    with open(os.path.join(path, 'explanation_metadata.json'), 'r') as f:
-        metadata = json.load(f)
-        is_global = metadata[ExplainType.GLOBAL]
-        is_local = metadata[ExplainType.LOCAL]
+    """Deserialize the explanation.
 
-    if is_local and is_global:
-        local_kwargs = _get_kwargs(path, shared_params + local_params)
-        local_explanation = _create_local_explanation(**local_kwargs)
-        global_kwargs = _get_kwargs(path, shared_params + global_params, local_explanation=local_explanation)
-        return _create_global_explanation(**global_kwargs)
-    elif is_local:
-        local_kwargs = _get_kwargs(path, shared_params + local_params)
-        return _create_local_explanation(**local_kwargs)
-    elif is_global:
-        global_kwargs = _get_kwargs(path, shared_params + global_params)
-        return _create_global_explanation(**global_kwargs)
+    Defined here for backwards compatibility.  The method has been moved to serialization.py.
+    Note this method is monkey patched there to avoid circular dependencies.
+
+    :param path: The path to the directory in which the explanation will be saved. By default, must be a new directory
+        to avoid overwriting any previous explanations. Set exist_ok to True to overrule this behavior.
+    :type path: str
+    :return: The deserialized explanation.
+    :rtype: Explanation
+    """
+    pass
